+++ programs/ibc-attestor/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod adapter_client;
 pub mod attestation;
+pub mod attestation_store;
 pub mod attestor;
 pub mod cli;
 pub mod header;
-pub mod height_store;
 pub mod server;
 pub mod signer;
